drivers: bluetooth: hci: silabs: fix radio IRQ priorities
Fix radio IRQ priorities. With the Controller update from SiSDK,
its behavior changed so that it overwrites the IRQ priorities.

Temporarily fix this by reconfiguring the priorities after the
Controller is initialized. This will be properly fixed in the
Controller later.

Also set the SYNTH IRQ priority on all HW, and set either the
RDMAILBOX or HOSTMAILBOX IRQ depending on which one exists.

Upstream-status: pending
Signed-off-by: Jori Rintahaka <jori.rintahaka@silabs.com>